refactor: update import paths for schemas to use the new lib structure
